게시글, 댓글 권한 설정, 예외처리

+++ src/main/java/com/example/velogclonebe/service/ArticleService.java
@@ -6,8 +6,10 @@ import com.example.velogclonebe.domain.dto.response.ArticleResponseDto;
 import com.example.velogclonebe.domain.dto.response.CommentGetResponseDto;
 import com.example.velogclonebe.domain.entity.Article;
 import com.example.velogclonebe.domain.entity.Comment;
+import com.example.velogclonebe.domain.entity.User;
 import com.example.velogclonebe.domain.repository.ArticleRepository;
 import com.example.velogclonebe.domain.repository.CommentRepository;
+import com.example.velogclonebe.domain.repository.UserRepository;
 import com.example.velogclonebe.exception.ApiRequestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,6 +24,7 @@ public class ArticleService {
 
     private final ArticleRepository articleRepository;
     private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
 
     // 게시글 리스트 조회
     @Transactional
@@ -32,24 +35,36 @@ public class ArticleService {
 
     // 게시글 작성
     @Transactional
-    public void setArticle(ArticleRequestDto articleRequestDto) {
-        Article article = new Article(articleRequestDto);
+    public void setArticle(ArticleRequestDto articleRequestDto, String username) {
+        Article article = new Article(articleRequestDto, username);
         articleRepository.save(article);
     }
 
     // 게시글 수정
     @Transactional
-    public void updateArticle(Long articleId, ArticleRequestDto articleRequestDto) {
+    public void updateArticle(Long articleId, ArticleRequestDto articleRequestDto, String username) {
+
         Article article = articleRepository.findById(articleId).orElseThrow(
-                ()-> new ApiRequestException("해당 글이 존재하지 않습니다."));
+                () -> new ApiRequestException("해당 글이 존재하지 않습니다."));
+
+        if (!article.getUsername().equals(username)) {
+            throw new ApiRequestException("수정 권한이 없습니다.");
+        }
+
         article.update(articleRequestDto);
     }
 
     // 게시글 삭제
     @Transactional
-    public void deleteArticle(Long articleId, ArticleRequestDto articleRequestDto) {
+    public void deleteArticle(Long articleId, String username) {
+
         Article article = articleRepository.findById(articleId).orElseThrow(
                 ()->new ApiRequestException("해당 글이 존재하지 않습니다."));
+
+        if (!article.getUsername().equals(username)) {
+            throw new ApiRequestException("삭제 권한이 없습니다.");
+        }
+
 
         // 글 삭제시 해당 글에 연관된 댓글들 삭제 (commentRepository 미반영)
         commentRepository.deleteAllByArticle(article);
@@ -58,14 +73,18 @@ public class ArticleService {
 
     // 게시글 상세페이지 요청 처리
     public ArticleResponseDto getArticleDetail(Long articleId) {
+
         Article article = articleRepository.findById(articleId).orElseThrow(
                 ()->new ApiRequestException("해당 글이 존재하지 않습니다."));
+
         // 게시글에 해당하는 댓글 리스트로 받아옴
         List<Comment> commentList = commentRepository.findAllByArticleOrderByCreatedAtDesc(article);
+
         // Dto로 넘겨주기 위한 작업
         List<CommentGetResponseDto> commentDtoList = commentList.stream()
                 .map(CommentGetResponseDto::new)
                 .collect(Collectors.toList());
+
         // 게시글 응답을 위해 articleresponsedto를 생성해서 반환
         ArticleResponseDto articleResponseDto = new ArticleResponseDto(article, commentDtoList);
         return articleResponseDto;
